Move serialization utilities to a separate file
