+++ frontend/components/Editor.js
@@ -4,6 +4,7 @@ import _ from "../imports/lodash.js"
 
 import { create_pluto_connection } from "../common/PlutoConnection.js"
 import { init_feedback } from "../common/Feedback.js"
+import { serialize_cells, deserialize_cells, deserialize_repl } from "../common/Serialization.js"
 
 import { FilePicker } from "./FilePicker.js"
 import { Preamble } from "./Preamble.js"
@@ -39,51 +40,6 @@ const uuidv4 = () =>
 /**
  * @typedef {import('../imports/immer').Patch} Patch
  * */
-
-/**
- * Serialize an array of cells into a string form (similar to the .jl file).
- *
- * Used for implementing clipboard functionality. This isn't in topological
- * order, so you won't necessarily be able to run it directly.
- *
- * @param {Array<CellInputData>} cells
- * @return {String}
- */
-function serialize_cells(cells) {
-    return cells.map((cell) => `# ╔═╡ ${cell.cell_id}\n` + cell.code + "\n").join("\n")
-}
-
-/**
- * Deserialize a Julia program or output from `serialize_cells`.
- *
- * If a Julia program, it will return a single String containing it. Otherwise,
- * it will split the string into cells based on the special delimiter.
- *
- * @param {String} serialized_cells
- * @return {Array<String>}
- */
-function deserialize_cells(serialized_cells) {
-    const segments = serialized_cells.replace(/\r\n/g, "\n").split(/# ╔═╡ \S+\n/)
-    return segments.map((s) => s.trim()).filter((s) => s !== "")
-}
-
-/**
- * Deserialize a Julia REPL session.
- *
- * It will split the string into cells based on the Julia prompt. Multiple
- * lines are detected based on indentation.
- *
- * @param {String} repl_session
- * @return {Array<String>}
- */
- function deserialize_repl(repl_session) {
-    const prompt = "julia> "
-    const segments = repl_session.replace(/\r\n/g, "\n").split(prompt)
-    const indent = " ".repeat(prompt.length);
-    return segments.map(function(s) {
-        return (indent + s).split("\n").filter((line) => line.startsWith(indent)).map((s) => s.replace(indent, "")).join("\n")
-    }).map((s) => s.trim()).filter((s) => s !== "")
-}
 
 const Main = ({ children }) => {
     const { handler } = useDropHandler()
@@ -1150,7 +1106,7 @@ patch: ${JSON.stringify(
         } />
                     <${FetchProgress} progress=${this.state.statefile_download_progress} />
                     <${Main}>
-                        <${Preamble} 
+                        <${Preamble}
                             last_update_time=${this.state.last_update_time}
                             any_code_differs=${status.code_differs}
                         />
@@ -1169,9 +1125,9 @@ patch: ${JSON.stringify(
                             }
                             disable_input=${!this.state.connected}
                         />
-                        <${DropRuler} 
+                        <${DropRuler}
                             actions=${this.actions}
-                            selected_cells=${this.state.selected_cells} 
+                            selected_cells=${this.state.selected_cells}
                             set_scroller=${(enabled) => {
                                 this.setState({ scroller: enabled })
                             }}
